+++ lib/monetize/core_extensions/nil_class.rb
@@ -1,0 +1,5 @@
+class NilClass
+  def to_money(currency = nil)
+    Money.new(nil, currency || Money.default_currency)
+  end
+end