+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
 ## master (next release)
+- Add NilClass extension
 - Added correct parsing of Brazilian Real $R symbol
 - Add testing task for  Brazilian Real parsing